# src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: Wallet Service - API
  version: 1.0.0
  contact:
    email: victorarduini@hotmail.com
servers:
  - url: http://localhost:8080
paths:
  /wallet:
    post:
      tags:
        - wallet
      summary: creates an user wallet
      operationId: createWallet
      parameters:
        - name: clientId
          in: header
          description: Client id of the wallet creation
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Create a new wallet to the specified user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRequest'
      responses:
        '200':
          description: Created Wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
  /wallet/{id}/order:
    post:
      tags:
        - wallet
      summary: Execute a client wallet order
      operationId: executeOrder
      parameters:
        - name: id
          in: path
          description: The wallet id
          required: true
          schema:
            type: integer
            format: int64
        - name: clientId
          in: header
          description: Client id of the wallet creation
          required: true
          schema:
            type: integer
            format: int64
        - name: walletIdTarget
          in: header
          description: Target wallet id of transfer, when type equal to TRANSFER
          required: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Create a new wallet to the specified user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletOrderRequest'
      responses:
        '202':
          description: Order will be processed
  /wallet/{id}/balance:
    get:
      tags:
        - wallet
      summary: creates an user wallet
      operationId: retrieveWalletBalance
      parameters:
        - name: date
          in: query
          description: Filter date to retrieve any historical balance related to this wallet
          required: false
          schema:
            type: string
            format: date
        - name: id
          in: path
          description: The wallet id
          required: true
          schema:
            type: integer
            format: int64
        - name: clientId
          in: header
          description: Client id of the wallet
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Created Wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'

  /user:
    post:
      tags:
        - user
      summary: creates an user
      operationId: createUser
      requestBody:
        description: Create a new user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Created User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

components:
  schemas:
    OrderType:
      type: string
      description: Transaction action applied
      enum:
        - WITHDRAW
        - DEPOSIT
        - TRANSFER
    
    WalletOrderRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/OrderType'
        value:
          type: number
    WalletRequest:
      type: object
      properties:
        name:
          type: string
          example: "My Wallet"
    WalletResponse:
      type: object
      properties:
        id: 
          type: integer
          format: int64
        name:
          type: string
          example: "My Wallet"
        balance:
          type: number 
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WalletBalanceResponse:
      type: object
      properties:
        balance:
          type: number
    UserRequest:
      type: object
      properties:
        name:
          type: string
          example: "Victor Arduini"
    UserResponse:
      type: object
      properties:
        id: 
          type: integer
          format: int64
        name:
          type: string
          example: "Victor Arduini"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Balance:
      type: object
      properties:
        balance:
          type: number
        createdAt:
          type: string
          format: date-time